# docs/swagger.yaml
basePath: /
definitions:
  ds.DraftInfoResponse:
    properties:
      draft_id:
        type: integer
      locations_cnt:
        type: integer
    type: object
  ds.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ds.GameResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      creator_name:
        type: string
      formed_at:
        type: string
      id:
        type: integer
      locations:
        items:
          $ref: '#/definitions/ds.LocationResponse'
        type: array
      moderator_name:
        type: string
      nickname:
        type: string
      random_pool:
        type: integer
      status:
        type: string
    type: object
  ds.GameStatus:
    enum:
    - черновик
    - удалён
    - сформирован
    - завершён
    - отклонён
    type: string
    x-enum-varnames:
    - GameStatusDraft
    - GameStatusDeleted
    - GameStatusFormed
    - GameStatusCompleted
    - GameStatusRejected
  ds.GamesListResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      creator_name:
        type: string
      formed_at:
        type: string
      id:
        type: integer
      moderator_name:
        type: string
      nickname:
        type: string
      random_pool:
        type: integer
      status:
        type: string
    type: object
...
  ds.LocationResponse:
    properties:
      id:
        type: integer
      is_randomed:
        type: boolean
      location_id:
        type: integer
      location_image:
        type: string
      location_name:
        type: string
      priority:
        type: integer
    type: object
...
  ds.MessageResponse:
    properties:
      message:
        type: string
    type: object
  ds.PlayersLocationGame:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      creator:
        $ref: '#/definitions/ds.User'
      creator_id:
        type: integer
      formed_at:
        type: string
      id:
        type: integer
      moderator:
        $ref: '#/definitions/ds.User'
      moderator_id:
        type: integer
      nickname:
        type: string
      status:
        $ref: '#/definitions/ds.GameStatus'
    type: object
...
  ds.User:
    properties:
      email:
        type: string
      id:
        type: integer
      is_moderator:
        type: boolean
      name:
        type: string
      password:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API для управления локациями и игровыми заявками
  title: Locations Project API
  version: "1.0"
paths:
  /api/games:
    get:
      consumes:
      - application/json
      description: Модераторы видят все заявки, обычные пользователи — только свои
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: Дата начала (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Дата конца (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.GamesListResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список заявок
      tags:
      - Games
  /api/games/{id}:
    delete:
      description: Мягкое удаление (смена статуса на 'удалён'). Доступ владельцу или
        модератору.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить заявку
      tags:
      - Games
    get:
      consumes:
      - application/json
      description: Возвращает детальную информацию о заявке с локациями. Доступ владельцу
        или модератору.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.GameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить заявку по ID
      tags:
      - Games
    put:
      consumes:
      - application/json
      description: Обновляет nickname заявки. Доступ владельцу или модератору.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ds.PlayersLocationGame'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить заявку
      tags:
      - Games
  /api/games/{id}/complete:
    put:
      consumes:
      - application/json
      description: Доступно только модераторам. approve=true завершает, false отклоняет.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: '{\'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Завершить или отклонить заявку
      tags:
      - Games
  /api/games/{id}/form:
    put:
      description: Переводит черновик в статус 'сформирован'. Доступ владельцу или
        модератору.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Сформировать заявку
      tags:
      - Games
  /api/games/cart:
    get:
      description: Возвращает ID черновика и количество локаций в нём.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.DraftInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Информация о черновике
      tags:
      - Games
...
  /api/gamloc/locations/{locationId}:
    post:
      description: Если черновика нет — создаёт новую заявку. Доступ владельцу или
        модератору.
      parameters:
      - description: ID локации
        in: path
        name: locationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ds.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Добавить локацию в заявку
      tags:
      - GamLoc
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"